web - cleanup factory
